Refactored device configuration assembly
